improved the design of the navigation bar and changed the default page to our taskPage
