+++ src/App.css
@@ -52,10 +52,13 @@
     background: "#00FF00";
 }
 
+.role-select-pos {
+    margin-left: 1100px;
+}
+
 .dropdown {
-    position: absolute;
-    top: 10px;
-    left: 10px;
+    position: relative;
+    right: 10px;
     display: inline-block;
 }
 
@@ -74,7 +77,7 @@
 }
 
 .welcome {
-    font-size: x-large;
+    font-size: xx-large;
     margin: 50px;
     color: rgb(71, 28, 3);
 }
@@ -169,10 +172,11 @@
 }
 /* App.css */
 /* Navbar.css */
+.navbar-taskP,
 .navbar {
-    position: absolute;
+    position: relative;
     top: 0;
-    right: 0;
+    left: 0;
     padding: 0.5rem;
     background: none;
     background-color: bisque;
