+++ src/pages/AboutPage.tsx
@@ -1,78 +1,12 @@
 import React from "react";
 import { Container } from "react-bootstrap";
-import { Link, Route, Routes } from "react-router-dom";
-import App from "../App";
-import TaskPage from "./taskPage";
-import { User } from "../interfaces/user";
-import { Task } from "../interfaces/task";
 
-interface AboutPageProps {
-    role: User;
-    setRole: (newRole: User) => void;
-    roles: User[];
-    setRoles: (newRoles: User[]) => void;
-    tasks: Task[];
-    setTasks: (newTasks: Task[]) => void;
-}
-
-export function AboutPage({
-    role,
-    setRole,
-    roles,
-    setRoles,
-    tasks,
-    setTasks
-}: AboutPageProps) {
+export function AboutPage() {
     return (
         <Container className="about">
-            <nav
-                className="navbar
-                        "
-            >
-                <ul>
-                    <li>
-                        <Link to="/about">About</Link>
-                    </li>
-                    <li>
-                        <Link to="/taskpage">Schedule Builder</Link>
-                    </li>
-                    <li>
-                        <Link to="/homepage">Homepage</Link>
-                    </li>
-                </ul>
-            </nav>
-            <Routes>
-                <Route
-                    path="/about"
-                    element={
-                        <AboutPage
-                            role={role}
-                            setRole={setRole}
-                            roles={roles}
-                            setRoles={setRoles}
-                            tasks={tasks}
-                            setTasks={setTasks}
-                        />
-                    }
-                />
-                <Route
-                    path="/taskpage"
-                    element={
-                        <TaskPage
-                            role={role}
-                            setRole={setRole}
-                            roles={roles}
-                            setRoles={setRoles}
-                            tasks={tasks}
-                            setTasks={setTasks}
-                        />
-                    }
-                ></Route>
-                <Route path="/homepage" element={<App />} />
-            </Routes>
             <h1>About Us</h1>
             <br></br>
-            <h3>Team Members</h3>
+            <h3>Team 18 - Members:</h3>
             <p>
                 Cornelia Meiss, Kaitlyn Sullivan,Aaron Oster, William Sharp,
                 Sydni Wright
@@ -83,7 +17,7 @@ export function AboutPage({
                 TimeWise is a planning based site that allows user to create
                 to-do lists and schedule their day. We provide a variety of
                 daily tasks and allow users to keep track of their daily
-                responsibilites and stay on track fro the day.{" "}
+                responsibilites and stay on track for the day.{" "}
             </p>
         </Container>
     );
